requirements.txt
pytest
python-box